requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
Deprecated==1.2.13
Django==4.1.2
idna==3.4
pycparser==2.21
PyGithub==1.56
PyJWT==2.5.0
PyNaCl==1.5.0
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
wrapt==1.14.1
parsel==1.6.0